사진과 함께 포스트 업로드 기능 프로토타입 완료.

+++ V-Album-Server/Domains/Photo/Photo.cs
@@ -8,11 +8,11 @@ public class Photo
     Guid? WorldUuid { get; }
     int Width { get; }
     int Height { get; }
-    int Size { get; }
+    long Size { get; }
     byte[] Hash { get; }
     string? Format { get; }
     
-    public Photo(Guid photoPhotoUuid, Guid photoPostUuid, int sortOrder, Guid? worldUuid, int width, int height, int size, byte[] photoHash, string? format)
+    public Photo(Guid photoPhotoUuid, Guid photoPostUuid, int sortOrder, Guid? worldUuid, int width, int height, long size, byte[] photoHash, string? format)
     {
         PhotoUuid = photoPhotoUuid;
         PostUuid = photoPostUuid;
